+++ actix-testing/CHANGES.md
@@ -1,5 +1,11 @@
 # Changes
 
+## [1.0.1] - 2020-05-19
+
+* Replace deprecated `net2` crate with `socket2`
+
+* Remove unused `futures` dependency
+
 ## [1.0.0] - 2019-12-11
 
 * Update actix-server to 1.0.0